pin/unpin

+++ lib/data/database/app_database.dart
@@ -24,7 +24,7 @@ class AppDatabase extends _$AppDatabase {
   AppDatabase([QueryExecutor? executor]) : super(executor ?? _openConnection());
 
   @override
-  int get schemaVersion => 12;
+  int get schemaVersion => 13;
 
   @override
   MigrationStrategy get migration => MigrationStrategy(
@@ -56,8 +56,8 @@ class AppDatabase extends _$AppDatabase {
         await migrator.alterTable(
           TableMigration(
             pinnedChats,
-            newColumns: [pinnedChats.type],
-            columnTransformer: {pinnedChats.type: const Constant('dm')},
+            // newColumns: [pinnedChats.type],
+            // columnTransformer: {pinnedChats.type: const Constant('dm')},
           ),
         );
       }
@@ -125,11 +125,24 @@ class AppDatabase extends _$AppDatabase {
           if (defaultOrganizationId != null) {
             await customStatement(
               'INSERT INTO pinned_chats '
-              '(id, folder_id, order_index, chat_id, pinned_at, type, organization_id) '
-              'SELECT id, folder_id, order_index, chat_id, pinned_at, type, $defaultOrganizationId '
+              '(id, folder_id, order_index, chat_id, pinned_at, organization_id) '
+              'SELECT id, folder_id, order_index, chat_id, pinned_at, $defaultOrganizationId '
               'FROM pinned_chats_old;',
             );
           }
+          await customStatement('DROP TABLE pinned_chats_old;');
+        });
+      }
+      if (from < 13) {
+        await transaction(() async {
+          await customStatement('ALTER TABLE pinned_chats RENAME TO pinned_chats_old;');
+          await migrator.createTable(pinnedChats);
+          await customStatement(
+            'INSERT INTO pinned_chats '
+            '(id, folder_id, order_index, chat_id, pinned_at, organization_id) '
+            'SELECT id, folder_id, order_index, chat_id, pinned_at, organization_id '
+            'FROM pinned_chats_old;',
+          );
           await customStatement('DROP TABLE pinned_chats_old;');
         });
       }
